Add helper classes for testing
GAETestCase is a TestCase base class where the developer can set the consistency_policy.  The class will then set the HRD consistency policy when running each test in the TestCase.

LiveServerTestCase runs dev_appserver in a separate thread for each test in the test case. This allows you to run Selenium tests against the live server.  Fixtures are properly supported.  However, the developer needs to be very careful to syncrhonize the testing code against the dev_appserver request handlers using the provided lock.