Add input reader utility classes for use with the mapreduce API. The main benefit of both is to be able to pass in a Django model instead of a db.Model to get Keys or entities.
Note that DjangoEntityInputReader returns a raw entity instead of a Django object instance.
This is particularly useful for working with old entities in the datastore that may not match the latest Django model.  Django's model classes could raise exceptions if properties were missing, but this class will allow fetching those entities.  The developer can then add the missing properties, or whatever.